requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.109.2
frozenlist==1.4.1
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
multidict==6.0.5
numpy==1.26.4
openai==1.11.1
portalocker==2.8.2
protobuf==4.25.2
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
qdrant-client==1.7.3
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
starlette==0.36.3
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
yarl==1.9.4
